+++ src/pages/profile/account-management/index.tsx
@@ -1,0 +1,5 @@
+import { ReturnComponent } from '@/shared'
+
+function AccountManagementPage(): ReturnComponent {
+  return <>AccountManagement</>
+}
