+++ src/pages/profile/general-information/index.tsx
@@ -1,0 +1,5 @@
+import { ReturnComponent } from '@/shared'
+
+function GeneralInformationPage(): ReturnComponent {
+  return <>GeneralInformation</>
+}
